Decouple pnAsyncCore from protocol details
pnAsyncCore[Exe] doesn't need to care if the right amount of initial
data was passed to AsyncSocketConnect. This also makes pnNetProtocol
independent of pnAsyncCore.

This uncovered a missing dependency from pnNetCli on pnAsyncCore
(resolved using a forward declaration).
