+++ Sources/Plasma/NucleusLib/pnAsyncCore/Private/pnAcIo.h
@@ -67,23 +67,6 @@ typedef struct AsyncCancelIdStruct *   AsyncCancelId;
 
 constexpr unsigned kAsyncSocketBufferSize   = 1460;
 
-/****************************************************************************
-*
-*   Socket connect packet
-*
-***/
-
-#pragma pack(push,1)
-struct AsyncSocketConnectPacket {
-    uint8_t     connType;
-    uint16_t    hdrBytes;
-    uint32_t    buildId;
-    uint32_t    buildType;
-    uint32_t    branchId;
-    plUUID      productId;
-};
-#pragma pack(pop)
-
 
 /****************************************************************************
 *
