+++ Sources/Plasma/NucleusLib/pnNetProtocol/Pch.h
@@ -52,7 +52,6 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 #include "HeadSpin.h"
 #include "pnNetBase/pnNetBase.h"
-#include "pnAsyncCore/pnAsyncCore.h"
 #include "pnNetCli/pnNetCli.h"
 #include "pnUUID/pnUUID.h"
